fix errors

+++ src/main/resources/db/migration/V4__Add_subscriptions_tables.sql
@@ -1,6 +1,6 @@
 create table user_subscriptions
 (
-    channel_id int8 not null references usr,
-    subscriber_id int8 not null references usr,
+    channel_id int4 not null references usr,
+    subscriber_id int4 not null references usr,
     primary key (channel_id, subscriber_id)
 );
